migrate monetary fields from `double` to `BigDecimal` for improved precision across `Invoice`, `Payment`, `RoomType`, and `Service` models, and update related calculations in `ReservationController`

+++ gs-serving-web-content-main/complete/src/main/java/com/duong/salesmanagement/controller/ReservationController.java
@@ -96,7 +96,7 @@ public class ReservationController {
             long nights = java.time.temporal.ChronoUnit.DAYS.between(
                     reservation.getCheckInDate(), reservation.getCheckOutDate());
             if (nights > 0) {
-                BigDecimal pricePerNight = BigDecimal.valueOf(room.getRoomType().getBasePrice());
+                BigDecimal pricePerNight = room.getRoomType().getBasePrice();
                 reservation.setTotalPrice(pricePerNight.multiply(BigDecimal.valueOf(nights)));
             }
         }
@@ -154,7 +154,7 @@ public class ReservationController {
                 long nights = java.time.temporal.ChronoUnit.DAYS.between(
                         reservation.getCheckInDate(), reservation.getCheckOutDate());
                 if (nights > 0) {
-                    BigDecimal pricePerNight = BigDecimal.valueOf(room.getRoomType().getBasePrice());
+                    BigDecimal pricePerNight = room.getRoomType().getBasePrice();
                     reservation.setTotalPrice(pricePerNight.multiply(BigDecimal.valueOf(nights)));
                 }
             }
